Unimos la rama dev a la master con todas las nuevas funcionalidades.

+++ database/migrations/2017_01_06_161342_create_claims_table.php
@@ -14,8 +14,8 @@ class CreateClaimsTable extends Migration
     public function up()
     {
         Schema::create('claims', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('rental_id')->unsigned();
+            $table->bigIncrements('id');
+            $table->bigInteger('rental_id')->unsigned();
             $table->enum('type', ['sugerencia', 'queja', 'reclamo'])->default('sugerencia');
             $table->string('title', 30);
             $table->string('description');
